sha1.c & sha2.c: Fix Static analysis bugs.
There were a number of variables which were (ostensibly for security
reasons) being cleared at function exit. However, these were not marked
as volatile so the compiler would be free to ignore this. They are now
set to be volatile to ensure that they get wiped.

Additionally, there were a few dead assignments that have been removed.